pyproject.toml
[project]
name = "memoh-lite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "milvus-lite>=2.5.1",
    "sentence-transformers>=5.2.3",
    "uvicorn>=0.41.0",
]
